pyproject.toml
[project]
name = "incident-tracker-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.121.0",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.44",
]
